+++ packages/maskbook/src/plugins/ITO/UI/CreateForm.tsx
@@ -18,7 +18,7 @@ import { useQualificationVerify } from '../hooks/useQualificationVerify'
 import { ITO_CONSTANTS } from '../constants'
 import { ExchangeTokenPanelGroup } from './ExchangeTokenPanelGroup'
 import { useCurrentIdentity } from '../../../components/DataSource/useActivatedUI'
-import type { PoolSettings } from '../hooks/useFillCallback'
+import type { PoolSettings } from '../hooks/useFill'
 import type { ExchangeTokenAndAmountState } from '../hooks/useExchangeTokenAmountstate'
 import ActionButton from '../../../extension/options-page/DashboardComponents/ActionButton'
 import { formatAmount, formatBalance, isGreaterThan, isZero } from '@dimensiondev/maskbook-shared'
